pyproject.toml
[tool.poetry]
name = "pingubot"
version = "0.0.5"
description = ""
authors = ["nootify"]

[tool.poetry.dependencies]
python = "^3.8"
python-dotenv = "^0.17.1"
"discord.py" = {extras = ["voice"], version = "^1.7.2"}
jishaku = "^1.20.0"
asyncpg = "^0.23.0"
gino = "^1.0.1"
wavelink = "^0.9.9"
psutil = "^5.8.0"
pytz = "^2021.1"
pytimeparse = "^1.1.8"

[tool.poetry.dev-dependencies]
coverage = "^5.5"
flake8 = "^3.9.2"
black = "^21.5b1"

[tool.black]
line-length = 120
include = '\.pyi?$'
exclude = '''
(
  /(
      \.eggs         # exclude a few common directories in the
    | \.git          # root of the project
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
  )/
  | foo.py           # also separately exclude a file named foo.py in
                     # the root of the project
)
'''

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
